+++ services/ParticipateService.py
@@ -1,0 +1,16 @@
+from flask import Flask, jsonify
+from flask import Blueprint
+
+from flask import Flask, request, jsonify
+from flask_restful import reqparse, abort, Api, Resource
+from flask_cors import CORS
+from contextlib import closing
+from config import config
+
+import connect_pg
+import psycopg2
+import requests
+import hashlib
+import json
+
+participate_app = Blueprint('participate_app', __name__)
